add recovery period logic

+++ src/wallet/TrueWallet.sol
@@ -204,14 +204,14 @@ contract TrueWallet is IAccount, Initializable, SocialRecovery, LogicUpgradeCont
     /// @notice Lets the owner set guardians and threshold for the wallet
     /// @param guardians List of guardians' addresses
     /// @param threshold Required number of guardians to confirm replacement
-    function addGuardianWithThreshold(address[] calldata guardians, uint256 threshold) external onlyOwner {
+    function addGuardianWithThreshold(address[] calldata guardians, uint16 threshold) external onlyOwner {
         SocialRecovery._addGuardianWithThreshold(guardians, threshold);
     }
 
     /// @notice Lets the owner revoke a guardian from the wallet and change threshold respectively
     /// @param guardian The guardian address to revoke
     /// @param threshold The new required number of guardians to confirm replacement
-    function revokeGuardianWithThreshold(address guardian, uint256 threshold) external onlyOwner {
+    function revokeGuardianWithThreshold(address guardian, uint16 threshold) external onlyOwner {
         SocialRecovery._revokeGuardianWithThreshold(guardian, threshold);
     }
